+++ bin/clean.js
@@ -1,4 +1,4 @@
-import Clean from './cleanClass.mjs';
+import Clean from './cleanClass.js';
 
 const clean = new Clean(process.argv);
 clean.run();
