+++ bin/common.js
