+++ bin/copyStatic.js
@@ -1,4 +1,4 @@
-import CopyStatic from './copyStaticClass.mjs';
+import CopyStatic from './copyStaticClass.js';
 
 const copyStatic = new CopyStatic();
 copyStatic.run();
