+++ bin/copyStaticClass.js
