+++ bin/packageExtension.js
@@ -1,4 +1,4 @@
-import PackageExtension from './packageExtensionClass.mjs';
+import PackageExtension from './packageExtensionClass.js';
 
 const packageExtension = new PackageExtension();
 packageExtension.run();
