+++ bin/postbuild.js
@@ -1,4 +1,4 @@
-import Postbuild from './postbuildClass.mjs';
+import Postbuild from './postbuildClass.js';
 
 const postbuild = new Postbuild();
 postbuild.run();
