+++ bin/prebuild.js
@@ -1,4 +1,4 @@
-import Prebuild from './prebuildClass.mjs';
+import Prebuild from './prebuildClass.js';
 
 const prebuild = new Prebuild(process.argv);
 prebuild.run();
