+++ bin/updateHelp.js
