+++ bin/watch.js
