+++ package.json
@@ -27,22 +27,22 @@
     "build:firefox:mv2": "npm run build --target=firefox-mv2",
     "build:firefox:mv3": "npm run build --target=firefox-mv3",
     "build:libs": "tsc -p ./src/script/lib/tsconfig.json",
-    "build:static": "node bin/copyStatic.mjs",
+    "build:static": "node bin/copyStatic.js",
     "build": "webpack --config bin/webpack.dev.js && npm run build:static",
-    "clean:all": "node bin/clean.mjs --all",
-    "clean:build": "node bin/clean.mjs --build",
-    "clean:dist": "node bin/clean.mjs --dist",
-    "clean": "node bin/clean.mjs",
-    "package": "node bin/packageExtension.mjs",
-    "postbuild": "node bin/postbuild.mjs",
+    "clean:all": "node bin/clean.js --all",
+    "clean:build": "node bin/clean.js --build",
+    "clean:dist": "node bin/clean.js --dist",
+    "clean": "node bin/clean.js",
+    "package": "node bin/packageExtension.js",
+    "postbuild": "node bin/postbuild.js",
     "postrelease:all": "npm run build",
-    "postrelease:build": "node bin/postbuild.mjs && npm run package",
-    "postrelease:firefox:mv2": "node bin/packageSource.mjs",
-    "prebuild": "npm run clean:build && node bin/prebuild.mjs --$npm_config_environment --$npm_config_target",
+    "postrelease:build": "node bin/postbuild.js && npm run package",
+    "postrelease:firefox:mv2": "node bin/packageSource.js",
+    "prebuild": "npm run clean:build && node bin/prebuild.js --$npm_config_environment --$npm_config_target",
     "prerelease:all": "npm run clean && npm run release:bookmarklet",
-    "prerelease:build": "npm run clean:build && node bin/prebuild.mjs --release --$npm_config_target",
+    "prerelease:build": "npm run clean:build && node bin/prebuild.js --release --$npm_config_target",
     "release:all": "npm run release:chrome:mv2 && npm run release:chrome:mv3 && npm run release:edgeLegacy && npm run release:firefox:mv2 && npm run release:firefox:mv3",
-    "release:bookmarklet": "node bin/prebuild.mjs --release --bookmarklet && webpack --config bin/webpack.bookmarklet.js",
+    "release:bookmarklet": "node bin/prebuild.js --release --bookmarklet && webpack --config bin/webpack.bookmarklet.js",
     "release:build": "webpack --config bin/webpack.prod.js && npm run build:static",
     "release:chrome:mv2": "npm run release:build --target=chrome-mv2",
     "release:chrome:mv3": "npm run release:build --target=chrome-mv3",
@@ -67,7 +67,7 @@
     "test": "npm run test:lint && npm run test:types && npm run test:addon && npm run test:cov",
     "type-check:watch": "npm run type-check -- --watch",
     "update:deps": "npx npm-check-updates --concurrency 10 --timeout 600000 --upgrade",
-    "update:help": "node bin/updateHelp.mjs",
+    "update:help": "node bin/updateHelp.js",
     "watch": "node bin/watch.js"
   },
   "devDependencies": {
